Prevent from converting label variables to numeric
Several label variable names end in `_LABEL`, including NAICS2017 codes. Suggesting to include this pattern to prevent the function to making it numeric, which results in NAs, especially in case of NAICS codes and labels. Similarly, attributes variables with symbols ends with `_F`. These variables also do not have to be converted to numeric.